First version of the API-Key Service has been proposed

+++ src/main/java/com/geopokrovskiy/repository/UserRepository.java
@@ -4,6 +4,6 @@ import com.geopokrovskiy.entity.UserEntity;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
 import reactor.core.publisher.Mono;
 
-public interface UserRepository extends R2dbcRepository<UserEntity, Integer> {
+public interface UserRepository extends R2dbcRepository<UserEntity, Long> {
     Mono<UserEntity> findByUsername(String username);
 }
