+++ src/main/resources/db/migration/V1__20231126.sql
@@ -2,7 +2,7 @@
 
 CREATE TABLE IF NOT EXISTS users
 (
-    id         INT PRIMARY KEY NOT NULL,
+    id         BIGINT PRIMARY KEY NOT NULL,
     username   VARCHAR(64)     NOT NULL UNIQUE,
     password   VARCHAR(64)     NOT NULL,
     status     VARCHAR(16)     NOT NULL,
